stock_ai/requirements.txt
# Stock AI Trading Framework Requirements

# Data & Analysis
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0

# Technical Analysis
ta>=0.11.0

# Visualization
matplotlib>=3.7.0

# Logging
loguru>=0.7.0

# Configuration
pyyaml>=6.0
pydantic>=2.0.0

# Broker API
alpaca-trade-api>=3.0.0

# Dashboard
streamlit>=1.28.0

# Optional: Database for results
# sqlalchemy>=2.0.0

# Optional: Advanced ML
# scikit-learn>=1.3.0
# torch>=2.0.0

